Summer Programs Description Page
HTML and CSS file for the Summer Programs Description Page.

Not using the header template at this point. Will be integrated next
time.

The page still need some work on cropping/resizing the main image.

The still of the fixed panel body and header is to be adjusted to the
design style.